requirements.txt
pdf2image==1.14.0
Pillow==8.1.2
pytesseract==0.3.7
python-decouple==3.4
